[dbal] add priority support on transport level.
